+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All Notable changes to the `chief` application template will be documented in this file. Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## unreleased
+
+## 0.2.8 - 2019-04-01
 - Added: A Page can now set a fixed base url segment (e.g. /news). This is set on the Page model via a `baseUrlSegment` property. This accepts a single string or an array of localized segments where the key should be the locale itself.
 - Added: pages.home route is not configurable in the chief config file.
 - Changed: minimum requirement for Chief is now php 7.2.*.